+++ src/coq_elpi_utils.ml
@@ -24,13 +24,11 @@ let err ?loc msg =
   let loc = Option.map to_coq_loc loc in
   CErrors.user_err ~hdr:"elpi" ?loc msg
 
-exception FatalElpiError of Pp.t
+exception LtacFail of int * Pp.t
 
-let () = CErrors.register_handler
-  (function FatalElpiError msg -> Some msg | _ -> None)
-let fatal_err ?loc msg =
+let ltac_fail_err ?loc n msg =
   let loc = Option.map to_coq_loc loc in
-  Loc.raise ?loc (FatalElpiError msg)
+  Loc.raise ?loc (LtacFail(n,msg))
 
 let feedback_fmt_write, feedback_fmt_flush =
   let b = Buffer.create 2014 in